requirements.txt
pandas>=2.0.0
pymysql>=1.0.2
python-dotenv>=0.19.0
openpyxl>=3.0.9
xlrd>=2.0.1 